minor refactor (#211)
* minor refactor

* lint
